requirements.txt
gunicorn
psycopg2-binary
django-bootstrap-form
django-bootstrap3-datetimepicker
xlrd
-r requirements-dev.txt